+++ src/components/index.ts
@@ -26,6 +26,8 @@ export * from './Checkbox'
 
 export * from './Chip'
 
+export * from './Code'
+
 export * from './ColorPicker'
 
 export * from './ColorPickerField'